requirements.txt
openai
requests
python-dotenv
gspread
oauth2client
